Rename the grammar module to grammar_ast.
